refactor: フォーマット系のディレクトリ構成をリファクタ
